\name{lines_RIGHT}
\alias{lines}
\alias{lines_RIGHT}
\title{Add Lines to a Plot}
\usage{
  lines_RIGHT(form, data, col = NULL, subset = NULL)
}
\arguments{
  \item{form}{y ~ x formula}

  \item{data}{data.frame object containing data}

  \item{col}{color of the points}

  \item{subset}{condition to subset data}
}
\description{
  lines_RIGHT draws a sequence of points in the current
  axis.
}

