\name{points_RIGHT}
\alias{points}
\alias{points_RIGHT}
\title{Add Points to a Plot}
\usage{
  points_RIGHT(form, data, col = NULL, subset = NULL,
    isString = FALSE)
}
\arguments{
  \item{form}{y ~ x formula}

  \item{data}{data.frame object containing data}

  \item{col}{color of the points}

  \item{subset}{condition to subset data}
}
\description{
  points_RIGHT draws a sequence of points in the current
  axis.
}

