\name{RIGHT}
\alias{RIGHT}
\title{Entry Function for RIGHT}
\usage{
  RIGHT(expr = { }, ...,
    title = "RIGHT: R Interactive Graphics via HTml",
    dir = tempfile(tmpdir = getwd()), isOverwrite = TRUE,
    supportRIGHT = getOption("supportRIGHT"))
}
\description{
  Entry Function for RIGHT
}

